some modified in sevral places

+++ src/app/services/serviceServices.ts
@@ -1,8 +1,31 @@
 import { Service } from "@prisma/client";
+import { jwtHelpers } from "../../helpers/jwtHelpers";
 import prisma from "../../libs/prisma";
 import ApiError from "../errors/apiErrors";
 
-const createService = async (data: Service): Promise<Service | null> => {
+import jwt, { JwtPayload } from "jsonwebtoken";
+
+const secret = process.env.JWT_SECRET;
+
+const createService = async (
+  token: string,
+  data: Service
+): Promise<Service | null> => {
+  if (!token) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const verify = jwtHelpers.verifyToken(token, secret as string);
+
+  if (!verify) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const decodeToken = jwt.decode(token) as JwtPayload;
+
+  if (decodeToken?.role !== "admin") {
+    throw new ApiError(401, "Unauthorized");
+  }
   const result = await prisma.service.create({
     data,
   });
@@ -29,6 +52,31 @@ const getSingleService = async (id: string): Promise<Service | null> => {
   return result;
 };
 
+const getAllServices = async (token: string): Promise<Service[] | null> => {
+  if (!token) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const verify = jwtHelpers.verifyToken(token, secret as string);
+
+  if (!verify) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const decodeToken = jwt.decode(token) as JwtPayload;
+
+  if (decodeToken?.role !== "admin") {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const result = await prisma.service.findMany({});
+  if (!result) {
+    throw new ApiError(401, "Bad Request");
+  }
+
+  return result;
+};
+
 const getAllAvailableServices = async (): Promise<Service[] | null> => {
   const result = await prisma.service.findMany({
     where: {
@@ -41,6 +89,7 @@ const getAllAvailableServices = async (): Promise<Service[] | null> => {
 
   return result;
 };
+
 const getAllUpcomingServices = async (): Promise<Service[] | null> => {
   const result = await prisma.service.findMany({
     where: {
@@ -55,9 +104,25 @@ const getAllUpcomingServices = async (): Promise<Service[] | null> => {
 };
 
 const updateSingleService = async (
+  token: string,
   id: string,
   data: Partial<Service>
 ): Promise<Service | null> => {
+  if (!token) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const verify = jwtHelpers.verifyToken(token, secret as string);
+
+  if (!verify) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const decodeToken = jwt.decode(token) as JwtPayload;
+
+  if (decodeToken?.role !== "admin") {
+    throw new ApiError(401, "Unauthorized");
+  }
   const result = await prisma.service.update({
     where: {
       id: id,
@@ -71,7 +136,25 @@ const updateSingleService = async (
   return result;
 };
 
-const deleteSingleService = async (id: string): Promise<Service | null> => {
+const deleteSingleService = async (
+  token: string,
+  id: string
+): Promise<Service | null> => {
+  if (!token) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const verify = jwtHelpers.verifyToken(token, secret as string);
+
+  if (!verify) {
+    throw new ApiError(401, "Unauthorized");
+  }
+
+  const decodeToken = jwt.decode(token) as JwtPayload;
+
+  if (decodeToken?.role !== "admin") {
+    throw new ApiError(401, "Unauthorized");
+  }
   const result = await prisma.service.delete({
     where: {
       id,
@@ -86,6 +169,7 @@ const deleteSingleService = async (id: string): Promise<Service | null> => {
 export const ServiceServices = {
   createService,
   getAllAvailableServices,
+  getAllServices,
   getAllUpcomingServices,
   getSingleService,
   updateSingleService,
